# clients/python/archiver-openapi.json
{
  "components": {
    "schemas": {
...
      "EnvelopeResponse": {
        "description": "Outer ``{\"detail\": ErrorEnvelope}`` wrapper for OpenAPI docs.\n\nPublic name (no leading underscore) so ``openapi-python-client`` generates a\ncleanly-named SDK model \u2014 the class name is what surfaces in\n``components/schemas`` and feeds the SDK code generator.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorEnvelope"
          }
        },
        "required": [
          "detail"
        ],
        "title": "EnvelopeResponse",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Unified error response body.",
        "properties": {
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional kind-specific structured payload (e.g. conflict id).",
            "title": "Data"
          },
          "errors": {
            "description": "Field-level problems; empty list when none apply.",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "kind": {
            "description": "Discriminator for client-side switching.",
            "enum": [
              "body",
              "schema",
              "domain",
              "lookup",
              "conflict",
              "auth",
              "unimplemented",
              "server"
            ],
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "description": "Human-readable summary; safe to surface to users.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "message"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "FieldError": {
        "additionalProperties": false,
        "description": "Single field-level validation problem.",
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional short machine-readable token (e.g. 'required').",
            "title": "Code"
          },
          "message": {
            "description": "Human-readable error message.",
            "title": "Message",
            "type": "string"
          },
          "path": {
            "description": "JSON-Pointer style path to the offending field.",
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "SourceRevisionCachePatch": {
        "additionalProperties": false,
        "description": "Request body for PATCH /source-revisions/{id}.\n\nBoth fields are optional (omitting leaves the DB column untouched).\nSupplying ``null`` explicitly clears the field.\nUse ``model_dump(exclude_unset=True)`` to distinguish omitted from null.",
        "properties": {
          "content_cache_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Cache Expires At"
          },
          "content_cache_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Cache Uri"
          }
        },
        "title": "SourceRevisionCachePatch",
        "type": "object"
      },
      "SourceRevisionCreate": {
        "description": "Request body for POST /source-revisions.\n\n``source_revision_id`` is optional and may be supplied by the client\n(e.g. Watcher) so the scratch file at ``content_cache_uri`` can be\nwritten under its final filename BEFORE the POST round-trips. When\nomitted, the server allocates a ULID. Idempotency on\n``(info_source_id, content_fingerprint)`` still wins on re-POST \u2014\na client-supplied ULID is honored on fresh inserts only; existing\nrows are returned as-is.",
        "properties": {
          "captured_at": {
            "format": "date-time",
            "title": "Captured At",
            "type": "string"
          },
          "content_cache_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Cache Expires At"
          },
          "content_cache_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Cache Uri"
          },
          "content_fingerprint": {
            "title": "Content Fingerprint",
            "type": "string"
          },
          "content_media_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Media Type"
          },
          "content_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Size Bytes"
          },
          "info_source_id": {
            "title": "Info Source Id",
            "type": "string"
          },
          "source_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Revision Id"
          }
        },
        "required": [
          "info_source_id",
          "content_fingerprint",
          "captured_at"
        ],
        "title": "SourceRevisionCreate",
        "type": "object"
      },
      "SourceRevisionOut": {
        "description": "Response body for source-revision endpoints.",
        "properties": {
          "captured_at": {
            "description": "UTC timestamp when the content was fetched and this revision recorded.",
            "format": "date-time",
            "title": "Captured At",
            "type": "string"
          },
          "content_cache_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "UTC timestamp after which the cached content at content_cache_uri expires.",
            "title": "Content Cache Expires At"
          },
          "content_cache_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Watcher's scratch-file URI for the cached fetch (e.g. a file:// path). Null after cache expiry or explicit clearance.",
            "title": "Content Cache Uri"
          },
          "content_fingerprint": {
            "description": "Content hash in 'sha256:<64 hex chars>' format. Together with info_source_id, forms the idempotency key.",
            "title": "Content Fingerprint",
            "type": "string"
          },
          "content_media_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "MIME type of the fetched content (e.g. 'text/html'), if recorded.",
            "title": "Content Media Type"
          },
          "content_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Size of the fetched content in bytes, if recorded.",
            "title": "Content Size Bytes"
          },
          "info_source_id": {
            "description": "ULID of the InfoSource this revision was captured from.",
            "title": "Info Source Id",
            "type": "string"
          },
          "source_revision_id": {
            "description": "ULID identifying this SourceRevision.",
            "title": "Source Revision Id",
            "type": "string"
          }
        },
        "required": [
          "source_revision_id",
          "info_source_id",
          "content_fingerprint",
          "captured_at",
          "content_size_bytes",
          "content_media_type",
          "content_cache_uri",
          "content_cache_expires_at"
        ],
        "title": "SourceRevisionOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "archiver",
    "version": "4.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/source-revisions": {
      "post": {
        "description": "Create or return an existing SourceRevision.\n\nIdempotent on ``(info_source_id, content_fingerprint)``. Returns **201**\non insert; **200** when the exact pair already exists.\n\nRaises:\n    404: ``info_source_id`` does not reference a known InfoSource.\n    422: ``content_fingerprint`` fails regex validation (Pydantic layer).",
        "operationId": "create_source_revision_api_v1_source_revisions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRevisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRevisionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Bad request (envelope)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Auth required (envelope)."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Forbidden (envelope)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Not found (envelope)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Conflict (envelope)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Validation failed (envelope)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Internal error (envelope)."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Source Revision",
        "tags": [
          "source-revisions"
        ]
      }
    },
    "/api/v1/source-revisions/{source_revision_id}": {
      "patch": {
        "description": "Partially update cache fields on an existing SourceRevision.\n\nOnly fields present in the request body are applied; omitted fields are\nleft untouched.  Sending ``null`` explicitly clears the field.\n\nRaises:\n    404: ``source_revision_id`` does not reference a known SourceRevision.\n    422: ``source_revision_id`` is not a valid ULID.",
        "operationId": "patch_source_revision_api_v1_source_revisions__source_revision_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "source_revision_id",
            "required": true,
            "schema": {
              "title": "Source Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRevisionCachePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRevisionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Bad request (envelope)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Auth required (envelope)."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Forbidden (envelope)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Not found (envelope)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Conflict (envelope)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Validation failed (envelope)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Internal error (envelope)."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Patch Source Revision",
        "tags": [
          "source-revisions"
        ]
      }
    },
...
  }
}